AI_DATA_ANALYST/requirements.txt
Flask==3.1.0
pandas==2.2.3
numpy==2.2.4
plotly==6.0.1
openpyxl==3.1.5
python-dotenv==1.0.1
google-genai
reportlab==4.2.5
gunicorn
